status-gateway/requirements.txt
pyserial
requests
prometheus-client